requirements.txt
pycryptodome==3.19.0
requests==2.26.0